drm/msm: Make ringbuffer could wrap
Currently, if read pointer is in front of write pointer,
when there is not enough space toward the end of the ringbuffer
for new commands, the driver should wait for the read pointer
reach to the end of ringbuffer.
This is not efficient, so remove the wait to end and write pointer
could back to the beginning of ring buffer without breaking the
commands list and catching up with read pointer.

Change-Id: I144249e2e517843b6bb468621ab41cbc9475867d
Signed-off-by: Kasin Li <donglil@codeaurora.org>
Signed-off-by: MOVZX <movzx@yahoo.com>