binder: fix race condition in new thread insertion to rb_tree
binder_get_thread() looks up a thread in the proc->threads rb_tree.
If not found, it drops the lock and allocates the new binder_thread.
The code then reacquires the lock and calculates the new insertion
point. Unfortunately, the search was not reinitialized so it used
the stale results from the first lookup. This works fine if no
other thread is created in-between. 1-line fix to reinitialize.

Change-Id: I236eaf2447ebe17119896b4002d89650666fe7b6
Bug: 32225111
Test: Tested manually
Signed-off-by: MOVZX <movzx@yahoo.com>