binder: serialize threads of each process
With the addition of fine-grained locking, the timing of some
transactions has changed and some apps have race conditions that
could never be hit before, but now can. This change adds a
per-proc mutex that provides serialization between all threads
of a proc (so it looks like the global lock case for threads in
a proc, but FGL between threads of different procs).

This can be modified later to allow a per-proc setting so apps
that are known to work with full FGL, can turn this off.

Bug: 34722727
Test: Tested manually
Change-Id: Idbc7dd9e813ecc237835f7e0d7462b420088a835
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: MOVZX <movzx@yahoo.com>