binder: fix race condition in death notification path
When releasing a node, we deliver death notification
to the requesting procs. There was a race condition
where the death notification is enqueued to the remote
proc and processed and kfree'd before sending the
wake_up signal. The crash occurred because we
deferenced "death" locally for the wake_up after
it had been kfree'd remotely.  Refactor to avoid
touching "death" once enqueued to the remote proc.

Bug: 35209998
Change-Id: Ic63707fd3c6f26f6d452846fad0c7d0b7bb6d94d
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: MOVZX <movzx@yahoo.com>