binder: ensure binder_node has reference when starting transaction
When setting up a transaction, we look up a binder_ref based on
the given handle and use the associated node as the target_node
for the transaction. However, since we call binder_put_ref on
the reference after discovering the node, there is a risk that
the ref will be deleted -- and if that is the last ref on the
underlying node, the node could also be deleted. When we
call binder_inc_node(), it may enqueue work for the proc. There
is a timing window where this could result in a corrupt list
entry on the proc->todo list when the node is kfree'd.

Bug: 35933974
Change-Id: Ie351c9e080a33b0cf0354f357876877f9d59784d
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>

Conflicts:
	drivers/staging/android/binder.c

Signed-off-by: MOVZX <movzx@yahoo.com>