binder: add missing locks for transaction_stack and return_error
A code review found a case where thread->transaction_stack is not
protected by the proc lock and should be serialized with
a todo list check. This can result in a thread waiting for
proc work when there is thread work to do. This has a workaround
in the code, which can now be reverted. Also found missing protection
for thread->return_error.

Added proc lock protection for transaction stack and for cases where
return_error needs to be serialized.

Bug: 36792545
Change-Id: I848ae994cc27b3cd083cff2dbd1071762784f4a3
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: MOVZX <movzx@yahoo.com>